# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PT. Kledo Berhati Nyaman",
        "description": "Back End Assessment V",
        "version": "1.0.0"
    },
    "paths": {
        "/api/approval-stages": {
            "get": {
                "tags": [
                    "Approval Stages"
                ],
                "summary": "Get list of approval stages",
                "operationId": "042d956a9fad73e84de890d02a0a497c",
                "responses": {
                    "200": {
                        "description": "List of approval stages"
                    }
                }
            },
            "post": {
                "tags": [
                    "Approval Stages"
                ],
                "summary": "Create a new approval stage",
                "operationId": "52989e6a14178a727bcc848e3acf8044",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "approver_id"
                                ],
                                "properties": {
                                    "approver_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Approval stage created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/approval-stages/{id}": {
            "get": {
                "tags": [
                    "Approval Stages"
                ],
                "summary": "Get approval stage by ID",
                "operationId": "137797aba6dbe900b5ea3f2f12554c74",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Approval stage found"
                    },
                    "404": {
                        "description": "Approval stage not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Approval Stages"
                ],
                "summary": "Update a approval stage",
                "operationId": "09e3ea8a8b87944defd865cd349851f9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "approver_id"
                                ],
                                "properties": {
                                    "approver_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Approval stage updated"
                    },
                    "404": {
                        "description": "Approval stage not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Approval Stages"
                ],
                "summary": "Delete a approval stage",
                "operationId": "7af9b0a6b398e3d4166fa6380f3239ba",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Approval stage deleted"
                    },
                    "404": {
                        "description": "Approval stage not found"
                    }
                }
            }
        },
        "/api/approvers": {
            "get": {
                "tags": [
                    "Approvers"
                ],
                "summary": "Get list of approvers",
                "operationId": "586a005e8bd9637c300f2b4990c7e37f",
                "responses": {
                    "200": {
                        "description": "List of approvers"
                    }
                }
            },
            "post": {
                "tags": [
                    "Approvers"
                ],
                "summary": "Create a new approver",
                "operationId": "8fddc9e935853028d340f48240c4f3ef",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Ana"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Approver created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/approvers/{id}": {
            "get": {
                "tags": [
                    "Approvers"
                ],
                "summary": "Get approver by ID",
                "operationId": "51e98d7bfb860d57eb5d5ec4ad6bb1c8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Approver found"
                    },
                    "404": {
                        "description": "Approver not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Approvers"
                ],
                "summary": "Update a approver",
                "operationId": "a87026aac3be4b0b6c6496da09b30351",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Budi"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Approver updated"
                    },
                    "404": {
                        "description": "Approver not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Approvers"
                ],
                "summary": "Delete a approver",
                "operationId": "cbf7a20cdcab3afd8e45b27ae07223dc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Approver deleted"
                    },
                    "404": {
                        "description": "Approver not found"
                    }
                }
            }
        },
        "/api/expenses": {
            "get": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Get list of expenses",
                "operationId": "9b702aa9314d471a9de3a121e5183c88",
                "responses": {
                    "200": {
                        "description": "List of expenses"
                    }
                }
            },
            "post": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Create a new expense",
                "operationId": "a764a04e6d047ab02a68f06c706e67f3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "example": "5100"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Expense created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/expenses/{id}": {
            "get": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Get expense by ID",
                "operationId": "97128dceb5c799772eef314ba1b9b045",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Expense found"
                    },
                    "404": {
                        "description": "Expense not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Update a expense",
                "operationId": "3132183b49b899f6dbc8e4c25b880461",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "example": "5000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Expense updated"
                    },
                    "404": {
                        "description": "Expense not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Delete a expense",
                "operationId": "68ebff62f3bfb2213565253975db2c52",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Expense deleted"
                    },
                    "404": {
                        "description": "Expense not found"
                    }
                }
            }
        },
        "/api/statuses": {
            "get": {
                "tags": [
                    "Statuses"
                ],
                "summary": "Get list of statuses",
                "operationId": "e96ea40c6f497546795df9ec612e1928",
                "responses": {
                    "200": {
                        "description": "List of statuses"
                    }
                }
            },
            "post": {
                "tags": [
                    "Statuses"
                ],
                "summary": "Create a new status",
                "operationId": "85739a0412fc324e86e6b45af20c7f0c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "MENUNGGU PERSETUJUAN"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Status created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/statuses/{id}": {
            "get": {
                "tags": [
                    "Statuses"
                ],
                "summary": "Get status by ID",
                "operationId": "ea554faff3dfcadd142b14010e8b3331",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status found"
                    },
                    "404": {
                        "description": "Status not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Statuses"
                ],
                "summary": "Update a status",
                "operationId": "7ba406b9e828f2ce3a0583d7c56583f6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Title"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated"
                    },
                    "404": {
                        "description": "Status not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Statuses"
                ],
                "summary": "Delete a status",
                "operationId": "4bcdd7d14d0707a62e62d2315c6149e7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status deleted"
                    },
                    "404": {
                        "description": "Status not found"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Approval Stages",
            "description": "Approval Stages"
        },
        {
            "name": "Approvers",
            "description": "Approvers"
        },
        {
            "name": "Expenses",
            "description": "Expenses"
        },
        {
            "name": "Statuses",
            "description": "Statuses"
        }
    ]
}